fix: llama_stack_api inspect API rename
when publishing llama_stack_api, `inspect.py` causes issues and gets confused to be the builtin stdlib inspect module.

This is due to the top level __init__.py we have. We need to rename inspect.py to inspect_api.py to avoid this conflict.

Also, uv sync
